Improve Jenkinsfile

+++ Jenkinsfile
@@ -40,8 +40,12 @@ pipeline {
             steps {
                 echo 'Deploy apk'
 
-                    writeFile(file: newApk, encoding: "UTF-8", text: readFile(file: apkLocation, encoding: "UTF-8"))
-                    echo 'Successfully renamed file'
+                script {
+                    if (fileExists(apkLocation)) {
+                        writeFile(file: newApk, encoding: "UTF-8", text: readFile(file: apkLocation, encoding: "UTF-8"))
+                        echo 'Successfully renamed file'
+                    }
+                }
             }
         }
         stage("Post Actions") {
